+++ Source/App/DecApp/EbMD5Utility.c
@@ -29,6 +29,8 @@
  * Includes
  ***************************************/
 #include <string.h>
+#include <assert.h>
+
 #include "EbFileUtils.h"
 #include "EbMD5Utility.h"
 
@@ -250,6 +252,10 @@ void print_md5(unsigned char digest[16]) {
 void write_md5(EbBufferHeaderType *recon_buffer, CLInput *cli, MD5Context *md5) {
     EbSvtIOFormat* img = (EbSvtIOFormat*)recon_buffer->p_buffer;
 
+    // Support only for 420 images
+    assert(cli->fmt == EB_YUV420);
+
+    const int bytes_per_sample = (cli->bit_depth == EB_EIGHT_BIT) ? 1 : 2;
     uint32_t y = 0;
     const uint8_t *buf = img->luma;
     uint32_t w = cli->width;
@@ -259,8 +265,8 @@ void write_md5(EbBufferHeaderType *recon_buffer, CLInput *cli, MD5Context *md5) 
 
     //luma MD5 generation
     for (y = 0; y < h; ++y) {
-        md5_update(md5, buf, w);
-        buf += stride;
+        md5_update(md5, buf, w * bytes_per_sample);
+        buf += (stride * bytes_per_sample);
     }
 
     w = w / 2;
@@ -271,15 +277,15 @@ void write_md5(EbBufferHeaderType *recon_buffer, CLInput *cli, MD5Context *md5) 
     //cb MD5 generation
     buf = img->cb;
     for (y = 0; y < h; ++y) {
-        md5_update(md5, buf, w);
-        buf += stride;
+        md5_update(md5, buf, w * bytes_per_sample);
+        buf += (stride * bytes_per_sample);
     }
 
     //cr MD5 generation
     buf = img->cr;
     stride = img->cr_stride;
     for (y = 0; y < h; ++y) {
-        md5_update(md5, buf, w);
-        buf += stride;
+        md5_update(md5, buf, w * bytes_per_sample);
+        buf += (stride * bytes_per_sample);
     }
 }