+++ Source/Lib/Decoder/Codec/EbDecIntraPrediction.h
@@ -25,7 +25,7 @@ void svt_av1_predict_intra(DecModCtxt *dec_mod_ctxt, PartitionInfo_t *part_info,
         EbBitDepthEnum bit_depth, int32_t blk_mi_col_off, int32_t blk_mi_row_off);
 
 void cfl_store_tx(PartitionInfo_t *xd, CflCtx *cfl_ctx, int row, int col, TxSize tx_size,
-    BlockSize  bsize, EbColorConfig *cc, uint8_t *dst_buff,
+    BlockSize  bsize, EbColorConfig *cc, void *pv_dst_buff,
     uint32_t dst_stride);
 
 #ifdef __cplusplus