+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Alt-ref pictures - temporal filtering
 - Adaptive Transform Block for INTRA
 - Adaptive QP scaling
+- Decoder - Support for Tiles and 10bit
 
 ## [0.5.0] - 2019-05-17